Signal gen 03

+++ audio_be/src/commonMain/kotlin/voices/strip/EnvelopeCalc.kt
@@ -8,6 +8,8 @@ import io.peekandpoke.klang.audio_be.voices.Voice
  * Calculates a single envelope value (0.0–1.0) at the given block position.
  * Uses the fixed release calculation: decays from the actual level at gate end,
  * not from sustainLevel.
+ *
+ * Long-to-Int conversion happens at the boundary; all arithmetic uses Int/Double.
  */
 fun calculateControlRateEnvelope(
     env: Voice.Envelope,
@@ -16,8 +18,8 @@ fun calculateControlRateEnvelope(
     gateEndFrame: Long,
 ): Double {
     val currentFrame = maxOf(blockStart, startFrame)
-    val absPos = currentFrame - startFrame
-    val gateEndPos = gateEndFrame - startFrame
+    val absPos = (currentFrame - startFrame).toInt()
+    val gateEndPos = (gateEndFrame - startFrame).toInt()
 
     val envValue = if (absPos >= gateEndPos) {
         val levelAtGateEnd = envelopeLevelAtPosition(env, gateEndPos)
@@ -32,7 +34,7 @@ fun calculateControlRateEnvelope(
 }
 
 /** Calculate the envelope level at a given position (attack/decay/sustain only). */
-fun envelopeLevelAtPosition(env: Voice.Envelope, absPos: Long): Double = when {
+fun envelopeLevelAtPosition(env: Voice.Envelope, absPos: Int): Double = when {
     absPos < env.attackFrames -> {
         val attRate = if (env.attackFrames > 0) 1.0 / env.attackFrames else 1.0
         absPos * attRate
